fix error

+++ core/src/xmake/sandbox/interactive.c
@@ -258,7 +258,7 @@ static tb_int_t xm_sandbox_loadline(lua_State* lua, tb_int_t top)
          * stack: arg1(sandbox_scope) scriptbuffer(top) -> ...
          * after: arg1(sandbox_scope) scriptbuffer scriptfunc(top) -> ...
          */
-        status = luaL_loadbuffer(lua, lua_tostring(lua, -1), lua_strlen(lua, -1), "=stdin");
+        status = luaL_loadbuffer(lua, lua_tostring(lua, -1), (size_t)lua_strlen(lua, -1), "=stdin");
 
         // complete?
         if (!xm_sandbox_incomplete(lua, status)) break;
